requirements.txt
slack_bolt
requests
python-dotenv
shodan
python-whois
filelock